# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es una API simple para realizar operaciones matemáticas básicas como suma, resta, multiplicación y división.",
        "title": "API de Operaciones Matemáticas",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/dividir": {
            "get": {
                "description": "Divide dos números",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Dividir",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Número a",
                        "name": "a",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Número b",
                        "name": "b",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "resultado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/multiplicar": {
            "get": {
                "description": "Multiplica dos números",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Multiplicar",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Número a",
                        "name": "a",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Número b",
                        "name": "b",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "resultado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/restar": {
            "get": {
                "description": "Resta dos números",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Restar",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Número a",
                        "name": "a",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Número b",
                        "name": "b",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "resultado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sumar": {
            "get": {
                "description": "Suma dos números",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sumar",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Número a",
                        "name": "a",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Número b",
                        "name": "b",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "resultado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}